substitute keras document aligner by sklearn extra trees classifier put dictionary based document aligner scripts in subpackage docalign use .gz compression for dictionary based alignment workflow
