+++ bitextor/Snakefile
@@ -1172,11 +1172,8 @@ rule bleualign:
 ### FILTERING AND CLEANING ######################################
 
 split_input_filename = "06_02.segalign"
-split_input_extension = ".gz"
-
 if SEGALIGN == "hunalign":
     split_input_filename = "hunalign.06_02.segalign"
-    split_input_extension = ".xz"
 
 
 # split segalign results into balanced chunks
@@ -1191,7 +1188,7 @@ checkpoint split_segalign:
     """
     input:
         lambda wildcards: [
-            f"{TRANSIENT}/{SRC_LANG}_{TRG_LANG}/{shard}/{SRC_LANG}{src_batch}_{TRG_LANG}{trg_batch}.{split_input_filename}{split_input_extension}"
+            f"{TRANSIENT}/{SRC_LANG}_{TRG_LANG}/{shard}/{SRC_LANG}{src_batch}_{TRG_LANG}{trg_batch}.{split_input_filename}.gz"
             for (shard, (src_batch, trg_batch)) in get_align_inputs(SRC_LANG, TRG_LANG)
         ],
     output:
@@ -1227,8 +1224,6 @@ checkpoint split_segalign:
                 CAT=cat
                 if [[ {input[0]} == *.gz ]]; then
                     CAT=zcat
-                elif [[ {input[0]} == *.xz ]]; then
-                    CAT=xzcat
                 fi
                 $CAT {input} \
                     | ( [ "{SRC_LANG}" = "{LANG1}" ] && cat || awk -F '\t' '{{ print $2,$1,$4,$3,$5 }}' OFS='\t' )\
