--- bitextor/data/model/Makefile.am
@@ -1,3 +1,0 @@
-modeldir = $(prefix)/share/bitextor/model
-
-model_DATA = keras.model keras.weights
