+++ requirements/requirements-dict-aligner.txt
@@ -1,2 +1,2 @@
 python-Levenshtein==0.12.2
-tensorflow==2.5.1
+scikit-learn==0.22.1
